Possible now to smooth out the back-reaction drag force that the dust particles exert on the gas.
Each particle is treated as a cube of the same size as a grid cell, but
centered on the particle. This way up to 8 grid cells are affected by the drag
force from one particle. The total influence of a particle still adds up to the
drag force in the case of no smoothing.

A particle sometimes changes the value of df in the first ghost zone. This
information is folded back into the main part of df after pde is done. This is
only implemented for one processor for now, so there is a dummy sub in
mpicomm.f90.