Added program structure for calculating the self gravity by solving the Poisson equation
  (d^2/dx^2 + d^2/dy^2 + d^2/dz^2)phi = 4*pi*G*rho

for the potential phi.

The potential is (will be) calculated before the pencil loop, and the gradient
of the potential is then added inside the loop.

The Poisson equation is (will be) solved in the Poisson moduke. Here different
ways of solving the partial differential equation will be added along the way.
The Fourier method already works in IDL, so will add that soon.