Possible to have fatal errors that occur only at a single processor. These are registred by calling fatal_error_local that is in the Messages module.
At the end of the time-step, the number of fatal errors is summed up and
communicated between the processors. The code dies if there are any errors.

Put in a local fatal error call if there is a NaN in dt1_advec. For some reason
a local crash of the simulation at one processor eventually let to dt=1e37,
which again caused the code to freeze when calculating the next tout. Not sure
if anyone else had such problems, but it was really annoying me.