When calculating the shock viscosity, make use of the code's ability to set boundary conditions on specific variables only (should make shock runs faster, but did not check).
Checked again that the shock viscosity is nice and periodic, even with shear.
Removed one unnecessary setting of the x-boundaries of ishock.

Stop the code if early_finalize is combined with shock. This does not set the
boundary conditions at the right time.

I hope this is okay, Tony. Couldn't resist applying the generalised bc routines
to the shock viscosity.