Implemented self-gravity on the dust particles. Had to make the self-gravity module a bit more general so that it does not itself access a specific variable in f, but rather receives the full right-hand-side of the Poisson equation. Thus gas and dust density can added and the result transferred to the Poisson solver.
The actual addition of dust density happens in the Particles_main module. Had
some trouble with circular dependencies, but was able to take the dependency on
Boundconds out of Particles_dust. It was used to set boundary conditions on the
dust number density for taking power spectra. Ahem... Or not... Fourier
transform don't really care very much about boundary conditions.