First attempt to export the neutron star business into special_nstar.f90
1. I have problems using the originally anticipated arrangement that
the new module should be called special_nstar.f90, because the compiler
produces src/nstar.o, so it is not underneath special anymore.
Can/should this be fixed at the Makefile level?

2. At this point I just removed a complete chunk of lines from hydro
and stuffed it into special_nstar.f90. This worked without problems.
Natalia, in start.in you just need to add
&special_init_pars
  leffective_gravity=T
 /
and remove this line from hydro_init_pars.

3. I guess all the special routines such as special_calc_hydro(f,df,p)
need also f as input. At the moment I've just done this for
special_calc_hydro, but of course we need it also in the other ones.