Made new module for self-gravity on the dust particles. The dust particles now have their mass spread out over the size of a grid cell, but centred on the particle. This allows us to resolve even tiny particle density pertubations that have not changed the number of particles in any grid cell, only the relative positions.
The gravity acceleration at the position of a particle is found by
interpolating the gradient of the potential from the nearest grid cells. This
required that the gradient is kept as an aux variable. Not a very pretty
solution, so perhaps one can interpolate and differentiate the potential in one
go, a possible future improvement.

Still a few issues (should be solved soon):

1. The smooth dust density does not work with shear yet.
2. The quadratic interpolation is still only in 2-D.