Die if a given choice of diffusivity (viscosity, resistivity, diffusion) does not have its corresponding coefficient set. I just spent most of a day debugging a crashing MHD run, and it turned out that the wrong resistivity coefficient had been set :-(.
As an exception, the basic diffusivity only warns if its coefficient is not
set, since this may often be the case in test problems.

I hope this does not cause trouble for anyone. One just has to remember to
comment out the diffusivity type as well as the coefficient if trying a run
with no diffusivity.