Also use cylindrical radius for angular velocity profile in 'keplerian' initial condition.
By the way, is there anyone using the 'mass_source' subroutine in density?
It's annotated as 'eos-merger', but I recall that Axel coded it
for me to use in the 2D globaldisk when I was just beginning.

I achieved better results using freezing, and never used this mass_source again.
It looks quite general, but is this subroutine obsolete?