1. added diagnostics to interstellar for heating and cooling rate (to check energy conservation, but I'm not convinced the result is ok). 2. in forcing and persist I just added some comments. 3. Tony, the interstellar module complains about unused pencils! We should remove them now, what do you think?
 pencil_consistency_check: checking the pencil case
pencil_consistency_check: OPTIMISATION POTENTIAL... pencil cp1tilde (   6)
is requested, but does not appear to be required!
pencil_consistency_check: OPTIMISATION POTENTIAL... pencil del2ss (  12)
is requested, but does not appear to be required!
pencil_consistency_check: OPTIMISATION POTENTIAL... pencil ee (  23)
is requested, but does not appear to be required!
pencil_consistency_check: OPTIMISATION POTENTIAL... pencil rho (  43)
is requested, but does not appear to be required!
pencil_consistency_check: OPTIMISATION POTENTIAL... pencil rho1 (  44)
is requested, but does not appear to be required!