Don't append $(FFLAGS_DOUBLE) to FFLAGS_GENERAL, just assign the value. This precents accumulation of multiple `-r8' flags in Makefile.inc.
B.t.w. does anybody recall what the name FFLAGS_GENERAL was supposed to
stand for (it is likely that I was the one who came up with it). The
`generality' may at some point have been the fact that it is used by F90
and F77, but by now FFLAGS are too. The name seems pretty unspecific and as
far as I can see, it is currently just used for the double-precision stuff.
  The contest for a better name is open...