Including particles_nbody.f90 as a replacement of particles_planet.f90
Just a small change of name to make it clear that what it does is
to assign a gravitational pull to the particles, letting them evolve
due to this mutual pull.

For few particles, using nbody is more efficient than solving the poisson
equation. Has to be generalized to npar particles before it can go to the
manual (currently just solves for two sink particles).