Created the sub 'fourier_shift_yz' that shifts an entire y-z plane by the amount shift_y in the y-direction. The interpolation is done in Fourier space to obtain maximum accuracy.
The non-MPI folding subroutines are already set to use Fourier shifting
(instead of 6th order interpolation). Does not seem to change the reference
data, so that is good.

Timing tests show that Fourier interpolation of one or two planes (as is needed
for the CIC and TSC particle assigment schemes) does not contribute noticably
to the wall clock time.

The parallel version of Fourier shifting still needs some thought...