Moved gravity_gas to particles_nbody.f90
As it evolves dfp (gravitational action of the gas onto the few particles),
the particles module is its natural place.

Purged wave_damping from planet.f90, along with loads of other minor changes.
Way too many and way too uninteresting to mention all of them. The plan is to
eventually move the whole planet code to the special module.