Add gas gravity to the dynamics of the planet particles. The particles should feel unsoftened newtonian gravity from every gas cell away from its accretion radius. The cells inside its accretion radius don't contribute to the acceleration.
The gas gravity produces non-cancelling forces on the particles, such that the
center of mass of the system of particles suffers a net acceleration. As I
don't want the center of mass dancing around the grid, in the end of each time
step a subroutin is called to reset the position of the center of mass
to the center of the grid.