Ok, this one here was a bit tricky, and I would like Anders and Tony opinion on it. I keep track of the sink particles in this simple way: the first nspar particles on the root processor.
These slots of fp and dfp - the first ones on the root processor - are spared for
these particles. The particle code makes two operations that change the indexing
of the particles:

 1. Redistribution of particles across processors due to boundary conditions
 2. Sorting of the particles in a pencil

If lparticles_nbody is activated, the redistribution does not affect the sink
particles, which means that they are not allowed to migrate (they are always kept
on the root processor).

The second is dealt with a simple unsorting of the sink particle. After the dust code
performs the sorting, I trace where the sink particles ended up, and bring them back
to the original position.