Added arrays fsp and dfsp, of dimension (nspar,mpvar). Their purpose is to store the sink particle properties, so no change in the abstraction of the dust particles module is needed.
The dust module deals with the particles as it always did (i removed the unsorting and the
non-allowance of migration for sink particles), but now it informs the nbody code about the
position of the sink particles, via fsp and dfsp

do k=1,nspar
   fsp(k,:) =  fp(ispar(k),:)
  dfsp(k,:) = dfp(ispar(k),:)
enddo

These arrays are then broadcasted to all processors. I intended to make them internal to the
nbody code, but as the communication is done on particles_main, I had to define them there. So,
unfortunately, the particle module has to know about the existance of these arrays even if
lparticles_nbody is switched off.

I also defined the indexing array ispar=ipar(1:nspar) in particles_cdata.f90, so
that I feed particles_nbody with

 integer dimension(nspar) :: ispar

instead of

 integer dimension(mpar_loc) :: ipar

But I don't know if it is a good thing to do since ipar is also defined in particle_cdata, thus
always available in the particle code. Shouldn't they both be defined in particles_main only?

I think I am getting infected by Tony's "to-hell-with-public-stuff" virus...