- added a new logical flag 'lwrite_2d' in 'init_pars' which allows to write the snapshots only in a (x,y) or (x,z) plane, ie the size of VAR* files are reduced by a factor 7 in the case of a pure 2D-run.
- made the corresponding changes in the IDL primary routines (pc_read_var,
pc_varcontent and pc_noghost) where one has to add the keyword '/run2D'
in the call of 'pc_read_var' to read 2D-snapshots. The reading is now
much more faster for 2D-runs!

PS: to ensure backward compatibility with older runs, I did not
automatically use the new 'param.lwrite_2d' value to check before reading
that the run is 2D or not, ie the abscence of the keyword '/run2d'
ensures the compatibility.