Corrected indentation, () around 'if' and 'while' added.
It took me an hour to even get something plotted, because the x-axis is not
defined per default, so tried to simplify things a bit. Give /xaxis, /yaxis or
/zaxis to state which axis goes on the x-axis of the plot, default is integers
from 1 to maximum of nx,ny,nz.

I hope this does break anyone's usage. Is anyone actually using this? It would
be great with some more comments in the code. It's still really, really hard to
read.