* Added default cases with calls to fatal_error in Sven's improvements     to the potential_point code.
To everybody:
When implementing IF or SELECT blocks always write code for all cases.
That includes the default or else case even if that code is only a call
to raise an error that the code shouldn't have gotten there. These
"fail-safes" are _very_ important in a muti-purpose multi-user code
like ours. If you see a missing case anywhere then do add it.

When a case is supposed to do nothing and it may be unclear that the
coder has recognised this fact then make it explicit by adding the
default case with a comment like  ! Do Nothing.   The compiler will
clean away any such empty blocks.