Fixed several compile and run-time problems in uu_driver(). Most of this probably has little effect on numbers, but replacing `n' (which was undefined and thus with most compilers potentially=0) by `n1' will somewhat change the quenching B-field.
The results have changed now, but only umax and shockmax are changed at
all, and the most notable change is in umax for t=0. So either the
initialization has changed, or the modifications in uu_driver mostly
affect the initial state:

   --it-------t------shockmax----umax--
<      0   0.00E+00  9.28E-02  7.59E-01
>      0   0.00E+00  9.09E-01  7.10E+00
<      5   4.79E-05  8.24E-03  1.26E-01
>      5   4.79E-05  1.21E-02  1.44E-01
<     10   9.57E-05  1.31E-02  2.34E-01
>     10   9.57E-05  1.62E-02  2.52E-01
<     15   1.44E-04  1.59E-02  3.35E-01
>     15   1.44E-04  2.04E-02  3.58E-01

Sven, can you make sense of the new numbers?