Moved totmass to density (previously in hydro, don't know why)
Renamed rhomphiz to rhomr, to keep consistency with the notation used for the cartesian averages, which is

qmxy = 2D  z-average of q  (depends on xy)
qmz  = 1D xy-average of q (depends on  z)

By the way, it is NOT the notation used in the 2D phi averages. They are phi averages, depending on r and z. They should be called "qmrz", not "qmphi".