Re-wrote default rule (compile-history etc) in structured form.
Axel, I really think this file should not be called xxx.local, as those
are the files that specify a run at a level above individual machines and
above, well, compile history. I would suggest any of:

  .compile-history
  compile.history

or, in order not to upset Tony,

  .build-history
  build.history

I have tentatively set the name to .build-history in this commit, but I am
not the one to decide.