Added header comment
+!  Note that to use lnTT as thermal variable, you may rather want to use
+!  entropy.f90 with pretend_lnTT=.true. As of March 2007, entropy.f90
+!  has way more options and features than temperature_idealgas.f90.

If somebody wants to add more advertising for temperature_idealgas to the
header, feel free to do so. But my understanding is that it currently does
only a small fraction of entropy.f90 (only one type of heat conduction,
etc.), so we should warn new users that there is more functionality
elsewhere.