Several routines changed at the same time:   M cdata.f90   M initcond.f90   M magnetic.f90   M param_io.f90   M register.f90   M sub.f90 But I will check them in separately, to avoid excessive logs.
The most important change is in bij_etc in sub.f90, as is best
seen by comparing the comment lines:
< !  calculate b_i,j = eps_ikl A_l,jk, as well as optionally,
> !  calculate b_i,j = eps_ikl A_l,kj, as well as optionally,
In cartesian it doesn't matter, but with spherical coordinates it does.
In the past I somehow felt the previous version nicer, but this is now
no longer a valid argument. I left in as a comment the old version, but
will delete it when Boris and Wlad have checked that the have not relied
on this previous strange version. (I will check the cylindrical stuff
myself later.)

There was another serious error:
<         bij(:,3,1)=bij(:,3,1)+aij(:,2,1)*r1_mn         +aa(:,2)*r2_mn
>         bij(:,3,1)=bij(:,3,1)+aij(:,2,1)*r1_mn         -aa(:,2)*r2_mn
I think the corresponding line in the cylindrical version has the right
sign already.

I also changed the indentation in a number of "call stop_it" commands.