Next part. Several routines changed at the same time:   M cdata.f90   M initcond.f90   M magnetic.f90   M param_io.f90   M register.f90   M sub.f90 But I will check them in separately, to avoid excessive logs.
Next change: in  cdata.f90 I added
<   integer :: ivisc_heat=0
>   integer :: ivisc_heat=0,ibb=0,ibx=0,iby=0,ibz=0,ijj=0,ijx=0,ijy=0,ijz=0

This is related to a useful debug method of outputting bb and jj in the
auxiliary array. I check in magnetic.f90 at the same time which has a
corresponding comment:
> !  Register an extra aux slot for bb if requested (so bb and jj are written
> !  to snapshots and can be easily analyzed later). For this to work you
> !  must reserve enough auxiliary workspace by setting, for example,
> !     SPECIAL   = special/maux6
> !  in your src/Makefile.local file, where special/maux6 is just a copy of
> !  nospecial.f90, but with the following line in the beginning.
> !     ! MAUX CONTRIBUTION 6

At the same time I also check in initcond.f90, because it contains a new
initial condition cosx_siny_cosz (used for tests only).