Committing a glimpse into the bright future when all diagnostic variables will be self-documenting. For that aim, the declaration of the idiag_XXX variables gets a comment that provides the LaTeX text for the table in Appendix H.3 of the manual:
  integer :: idiag_urms=0       ! DIAG_DOC: $\left<\uv^2\right>^{1/2}$
  integer :: idiag_umax=0       ! DIAG_DOC: $\max(|\uv|)$
  integer :: idiag_ruxm=0       ! DIAG_DOC: $\left<\varrho u_x\right>$
                                ! DIAG_DOC:   \quad(mean $x$-momentum density)

extract-diag-doc (the script that extracts the comments and writes them in
LaTeX format to a file) seems to basically work, but I haven't hooked this up
to manual.tex yet.

Note: In hydro.f90: found a number of diagnostic variables that were not
being properly reset in rprint_hydro. It is highly likely that there are more.
  It would be desirable to automate some of this, such that instead of
ihaving to put the two lines

  integer :: idiag_u2m=0        ! DIAG_DOC: $\left<\uv^2\right>$
  integer :: idiag_urms=0       ! DIAG_DOC: $\left<\uv^2\right>^{1/2}$
  ! [...]

  ! [Then, in rprint_hydro:]
  idiag_u2m=0
  idiag_urms=0

in hydro.f90, we would just put

  ! BEGIN DIAGNOSTIC
  !   u2m    $\left<\uv^2\right>$
  !   urms   $\left<\uv^2\right>^{1/2}$
  ! END DIAGNOSTIC

there and have the code generated automatically. Unfortunately, I cannot
come up with a reaonable way of doing this without one or two additional
include files for each physics module :-(.