Caught a subtle error that occurs if one or more processor does not calculate a given diagnostic variable. Then itype_name will be zero at those procs, but possibly not zero at all the other procs. In the subroutine 'diagnostic' there will be a different count of nmax_count and nsum_count on each processor, leading to a cryptical MPI error message when trying to reduce the max and the sum of arrays of conflicting sizes ("receive buffer of size 56 is too small for send buffer of size 64" or something similar).
Fixed this for the drag force so that the diagnostic of the drag force
time-step is calculated even if a pencil has no particles. I wonder if there is
a better way to ensure that the diagnostics type is set properly on all
processors. Ultimately it would be good to have diagnostics in a structure or
something and then do all set up of types etc. before the main time loop.