Generalized what was done with the spheromak globals by allowing for external B, J and E fields as globals in the f array. They are added component wise, so it is possible to only set, say, B_z and E_x, and not spend memory on components that are zero.
Is the global_spheromak module actually obsolete?

I think we really needed a general way to have imposed magnetic fields in the
code, since the vector potential formulation makes net flux field
initial conditions difficult to implement (A is not periodic, so boundary
conditions are tricky).

I'll try to improve on this in the next days. So far I managed to make a
hydrostatic-magnetic equilibrium with a net B_y field (thus J_x, whereas the
extra E comes automatically from B_y).