Sorry, something ate my login message. This should (at least roughly) fix the problem with nogsl.c on some machines.
I am also removing gsl_wrapper_r8.c, which shouldn't be necessary by now.
This might require a bit of support in the Makefile, so Dhruba please let
me know if there are problems.

And no, I didn't have time to test thoroughly. And yes, this is not good
style...