Ok, I think I fixed the upwinding in cylindrical coordinates. The der6 subroutine had a df=df*rcyl_mn1**6 that I added some time ago. This was not compatible with the formulation of the upwinding term, which is
 |u_j|*alpha*(dx_j)^5*d6(f)/dx_j^6

with dx_j=(r*dphi), this is

 [|u_phi|/r] * {alpha* d6(f)/dphi}

The subroutine u_dot_grad_scl takes the output of der6 (in curly  brackets)
and multiplies by what is in the square brackets.

The output of the subroutine der6 HAS WRONG DIMENSIONS unless upwinding is
used. The correct formulation would be having the rcyl_mn1**6 factor in der6
and multiplying by rcyl_mn**5 in u_dot_grad_scl, but that's redundant use of
powers.

Added a fatal_error if der6 or der6_other is called without upwinding.

Spherical coordinates were already correct.