Okay, I finally managed to code hyperdissipation for cylindrical and spherical coordinates
 fdiss = chi^(3)_j dot (d^6/dxj^6) q          (1)

where chi^(3) is the third order dissipation coefficient (diffusion, viscosity or resistivity), xj are the coordinates (r,theta,phi) or (pomega,phi,z). q is the quantity to the dissipated (density, the three components of velocity or magnetic potential). The derivatives are in non-coordinate basis and summation over repeated indices applies.

This is for cells of different aspect ratio, so chi^(3)_j is a tensor (1st rank in this case). However, it scales with the length elements dxj, so that this third order dissipation can be written in terms of the first order scalar chi and the respective infinitesimal

 chi^(3)_j = chi (dxj/pi)**4                  (2)

Substituting (2) in (1) we arrive at,

 fdiss = chi/pi**4 * d^6(q) * (1/dxj**2)      (3)

Eq.(3) is coordinate independent, and therefore works also for Cartesian coordinates, being therefore another (general) way of expressing the anisotropic hyperdissipation I coded once. The beauty of this formulation is that it scales with the resolution, so you set the hyperviscosity like if you were setting the coefficient for the second order Laplacian version (like, 1e-3 instead of 1e-15).
Careful though: THIS IS JUST A HIGH FREQUENCY FILTER. It's conserving neither momentum nor mass. It does conserve these quantities in the Cartesian case, but that's all (even then, it is not conserving angular momentum in the grid scale).

Please check and see. Feedback mostly appreciated.
-Wlad