Thanks Steve!
Using "rho" for the cylindrical distance is dangerous... can lead one
to code p%lnrho*p%uu(:,2) when thinking about rcyl_mn*p%uu(:,2) :p

Anyhow, uu(:,2) in cylindrical coordinates is already dimensionalized (=r\dot\phi) so the formulation of the Coriolis force for cylindrical and cartesian is the same.

But it's better to have it in a separated subroutine because precession is
not implemented for cylindrical.