Reinstated run2D option. Sorry about that Boris! I see that it is very practical to be able to write 2-D data without ghost zones in the missing direction.
Please test and see if it works.

One problem was of course that run2D is not documented in the files (it is in
the manual, I realised later). I would encourage everyone (including myself) to
be more generous with comments, especially for parts of the IDL files whose
functionality is not obvious.