Possible to get magnetic tension and magnetic pressure with
  pc_read_var, obj=ff, variables=['mten','mpres'], /magic

Both terms are calculated using b_i,j, calculated with derijcurl(aa).

Implemented dependencies of magic variables. There is a list in the beginning
of the file. This will e.g. calculate bb before mten and insert bij in the
variable list if it is not already there.

Dependencies are easily specified through commands of the type:

  pc_magic_var_dep, variables, tags, 'mten', 'bb'