Sum the gravity of all massive particles, map it to the grid and interpolate to (dust) particle positions. Factor 4 speed up for 1e6 particles with 10 sinks as compared to the other way of computing it (loop particle-particle).
It is not so much, though, since most of the time (~80% for 2D 256^2 resolution run) is spent in sort_particles_imn. Moreover, when weighting speed x accuracy,  the long loop loses in the former but gains in the latter.

Anyway, having a switch for both ways of calculating it might come in handy someday.