Implemented removal of particles.
 bcpx='rmv'

The particle will be removed from the simulation if crossing the x boundary.
The structure is similar to Anders' migrating business.  Looping backwards
from k=npar_loc to k=1, copy the last particle in the processor to the slot
left by the excluded particle

 fp(k,:)=fp(npar_loc,:)
 if (present(dfp)) dfp(k,:)=dfp(npar_loc,:)

And reduce the number of particles in the processor by one

 npar_loc=npar_loc-1

The index ipar(k) of the excluded particle is written to the file
data/procN/rmv_par.dat so pvar.dat knows which particles were lost
(thanks Anders for the hint).