Add subroutine for dynamical allocation of sink particles
It tests the grid for cells whose resolution exceeds the Jeans length
(pi*cs2/(G*rho)) in either the gas or the dust phase

    rho > rho_J = pi*cs2/(G*Delta)

    Delta=min(dx,dy,dz)

If the criterion is violated in the gas phase, set the gas density to the jeans value and collapse the excess mass (rho-rho_jeans)*dvolume into a sink particle.

For the dust phase, replace the sound speed by the local rms of the particle's velocity. If the criterion is violated, remove all particles of the cell and replace them by a sink particle at the center, with the mass and momentum of the combined dust clump that collapsed.