Fixed the dependencies of testperturb. Now the code compiles. The steps were (1) In Makefile.depend defined
 TESTPERTURB_OBJ=$(TESTPERTURB).o
 TESTPERTURB_SRC=$(TESTPERTURB).f90


(2) made param_io and register depend on it, so the rprint_list and read/write parameters of testperturb are seen. (3) As it used rk_2n, it must depend on time-step.

On Makefile.src, (4) remove testperturb from the physics variables, otherwise it gets compiled too early through the dependence of equ on physics. (5) Both start and run have to depend on testperturb since run needs to know about testperturbing and start needs to know the registering stuff of step (2).

The dependence of start on testperturb made start depend also on equ and timestep, the dependencies of testperturb. That is undesirable, but not a big problem, I suppose.

Axel, can you check if it is working?