Implemented curl and div for spherical coordinates
Also, changed the derivatives to non-coordinate basis, as it is in the code itself. That, however, forces zder, yder and all files having curvilinear corrections to call pc_read_param, in order to get the param.coord_system string that lets the idl code know if we are in cartesian, cylindrical or spherical.