Found a crazy thing concerning equ.f90: using the pgprof profiler in connection with pgf90, I reduced the computing time by about 60% on my sample! see
http://www.ast.obs-mip.fr/users/dintrans/tmp/pencil/index.html

The pb was that many dummy subroutines were called in equ.f90 because
their associated flag were unused in 'equ.f90', eg,

before: call dlncc_dt(f,df,p)
now: if (lpscalar) call dlncc_dt(f,df,p)

and so on with the magnetic, dust evolution or neutral modules. In
addition, the magnetic pencils were computed even if the run is a
pure-hydro one.

Question for Anders: in the dust evolution stuff, if ldustvelocity=T, does
it also mean that ldustdensity=T? If positive, we can maybe re-arrange
a bit further equ.f90 by

!
!  Dust evolution
!
if (ldustvelocity) then
  call duud_dt(f,df,p)
  call dndmd_dt(f,df,p)
endif

It remains something to do concerning the gravity_simple module because
it takes too much time (about 12% of the run)...

Boris.