Major cylindrical bug: p%aa was not been requested, so p%bb was wrong!
The call was

      if (lpencil_in(i_bb)) then
        if (lspherical_coords) lpencil_in(i_aa)=.true.
        lpencil_in(i_aij)=.true.
      endif

So I changed lspherical_coords to .not.lcartesian_coords. Without this, the Aphi/rcyl_mn curvilinear correction
term in the vertical component of the curl was not being taken into account.

It illustrates again the need for abundant calls to fatal_error for cylindrical when implementing spherical, and vice-versa.

Of course, I could have switched lpencil_check on and caught this bug way earlier....