Implemented general drag law for generic Knudsen and Mach numbers
   Fdrag = [3*Kn/(3*Kn+1)]**2 * Feps +  [1/(3*Kn+1)]**2 * Fsto

Where Kn is the Knudsen number, Feps is Epstein drag and Fsto the Stokes drag

   Feps=-sqrt(128*pi)/3*a**2*rhog*cs*fd*Delta(u)

   Fsto=-6*pi*a*kd*mu_kin*Delta(u)

In the Epstein drag law, fd is the correction fact for arbitrary mach numbers

   fd=sqrt(1 + 9*pi/128*Mach**2)

And in the Stokes drag, mu_kin is the kinematic viscosity of the gas

      mu_kin=1/3*rhog*vth*lambda

and vth=sqrt(8/pi)*cs is the mean thermal velocity of the gas. kd is a factor
that contains the Reynolds number of the flow over the particle.

The implementation works for variable density, sound speed and velocity in the 3D case. For 2D, it assumes that the particle's internal density and the reference Keplerian speed are equal to 1.