Another try at my `warn me if precision has changed' patch. Solved the previous problem by including src/Makefile.local from Maekfile(.parent), too.
If again a number of auto-tests mysteriously breaks, I am the likely
culprit. The problem last time was that we compiled for single precision
(by mistake), but still used mpicomm_double, which meant that some
MPI routines operated on double precision numbers when the calling routine
thought it just gave them access to singles.
  I think the compiler could spot this if MPI was a Fortran module. Do
we have any test system (that could do periodic tests) where this is the
case?