Improved the pencil declaration scheme. Dimensions are now no longer specified in mkcparam, but rather directly in the header. For example:
! PENCILS PROVIDED aa(3); a2; aij(3,3); bb(3); bbb(3); ab; uxb(3)

In this line there are both scalar, vector and matrix pencils. More complicated
pencils, involving ndustspec and nchemspec, are defined in a similar way.

This should make pencil declaration a bit easier and more transparent.

Notice that the pencils are separated with ";" now, because the "," is used in
the dimensionality declaration.

The auto-test passed fine, except for solar_system, no-modules and
heating_ionize, but those errors seem to be unrelated to the pencil case.