Characteristic boundary conditions: non-reflecting outlet and inlet, has been added.
These boundary conditions are imposed by solving a so-called characteristic
form of the fluid equations on the boundaries, as opposed to setting
actual values of the variables in the f-array. The method is called
Navier-Stokes characteristic boundary conditions (NSCBC).

Because this way of treating boundaries could not be done using the existing
structure for boundary conditions (using the bcx1, bcx2 variable etc.), some
new variables and subroutines had to be added to the code.

The treatment is called from equ.f90 right after the yz-loop. Relevant
variables and subroutines have the short-hand nscbc in their names.

grep nscbc * | wc -l
54

In addition, a sub doing a 4th order one-sided derivative was added to
deriv.f90

The current implementation only solves a simplified version of the
equations, namely a set of so-called local one-dimensional inviscid
(LODI) relations. This means that transversal and viscous terms are
dropped on the boundaries.