Possible to *not* precalculate the interpolated gas velocity, by setting lnostore_uu. This gives a speed gain of around 5% - probably because the precalculation of the gas velocity involves an allocatable array that adapts its size to the number of particles in the considered pencil.
There is also a worry that the interp_uu array could become too big if all
particles cluster in a single pencil, or in a single grid point. In that case
one can set lnostore_uu=T in &particles_run_pars. But there should be a major
advantage in precalculation already if the interpolated gas velocity is needed
twice.

Also removed some experimental gravity profile that I had checked in by mistake.