allow 2 ways to visualize the spherical slices
  1. plain patching of the meridional planes using the regular grid
  
  2. interpolate the meridional planes to cartesian coordinates 
     and mask the radii and theta out of range (default)

The 2nd reproduces the curvature, so it is more appropriate for simulations 
with large latitude coverage. The slices might get big, tough, since the 
grid interpolation refines the resolution to get decent sampling when 
transforming from theta to z. Calling sequence

  rvid_box_sph,'lnrho',min=0,max=2

The 1st one doesn't reproduce the curvature of the meridional plane, but it 
looks virtually identical to the 2nd for small latitude coverage. Calling 
sequence

  rvid_box_sph,'lnrho',min=0,max=2,/nointerpz