Fetch particle values (fp) to the special module.
special/localshock.f90 defines a mask (of radius < rmask) 
around a massive particle, where the shock dissipation 
(diffusion, viscosity, etc...) is enhanced.

This allows a simulation with massive particles (that attract A LOT 
of gas) to use more moderate (or even vanishing) shock viscosity in 
the more quiescent regions of the mesh.