Inconsistency in the particle migration subroutine
A particle domain is from (AND INCLUDING) the lower processor boundary 
to up to (and excluding) the upper processor boundary. Therefore a 
particle needs to be migrated if 

	(fp(k,iyp) < y0) or (fp(k,iyp) >=y1)

	(fp(k,izp) < z0) or (fp(k,izp) >=z1)

I.e., the equal sign was missing. 

My star and my planets, which I usually assign to position y=0, landed 
exactly at a processor boundary, and therefore were not migrated. The ipar 
got wrong and the star ended up with zero mass. 

This thing was sitting there since August (revision 9716). It's not a bug, 
it's a trilobite!

I didn't notice it before because the auto-test is not capturing this 
behaviour. The globaldisc sample is serial, dust_globaldisk doesn't 
use n-body, and solar-system is in cylindrical coordinates, where the 
star is not at y=0 (but at y=-!pi) at start time. The planets, that are at 
y=0, move towards positive values of phi, so they don't migrate initially.