Added a --time-limit option that uses reaper to kill run.x processes after a given time.
We can now e.g. do
    auto-test --time-limit=2m
to limit the run.csh phase of each test to 2 minuts maximum. If run.x
processes run longer than the limit, they will get killed (and the
corresponding auto-test failes).
    pencil-test -t 2m [...]
or
    pencil-test --time-limit 2m [...]
does the analogous thing for pencil-code.

The plan is now to run all tests in the samples/ hierarchy and fix or
shorten any test that isn't successfull.