Moved freezing to after the imn loop. df may still be modified outside of the considered imn pencil (e.g. by particle drag force), so one has to do freezing after all contributions to df have been calculated.
Wlad, I also changed the coding so that if wfreeze_int or wfreeze_ext are zero,
then a simple step function is used, rather than calling quintic_step. This
should also work when border profiles are used (if the width is set to zero).

Unfortunately r_mn and rcyl_mn need to be recalculated now. It must still be
done by the Grid module, so simply call calc_pencils_grid again. There may be a
better, faster solution.