Adapted xz-averages (x-dependent) to nprocx>1. Also set lwrite_ic=T by default.
Let Sub use Cdata and Cparam globally, to avoid specifying in every single
subroutine what we are using. I think the latter is good practice for
subroutines, but excessive for parameters from Cdata and Cparam which are
supposed to be global anyway (so no need to feel bad for using them!).