Managed to merge Print and Diagnostics. This makes it a lot easier to overview what subs are used for calculation of averages.
Axel, I did not know what to do with the calc_mXXXX stuff. I think it belongs
in equ.f90, so I copied it there (commented out for now).

We could still consider renaming Diagnostics -> Prints, for historical reasons
and because I think the idea of the Print module was exactly to collect
everything in one place.