One of Anders' last commits deleted Axel's "tanhy" initial condition and my "initial-condition" border profile in density.f90. Re-introducing them here.
There's something weird with svn. It doesn't accuse conflicts as nicely as
cvs. This thing has already happened before, when a commit of Natalia
deleted a whole bunch of my coding. The bad thing is that it goes
unnoticed unless someone's runs start crashing (like the samples
globaldisk and dust-turb-globaldisk, which failed because the border
profile wasn't there anymore).