Fixed problem that this script had. The last subroutine coded in nospecial.f90 carried the line containing "include special_dummies.inc" into the special_dummies.inc file. It naturally caused an infinite loop.
The problem was that the flag for writing a line into special_dummies.inc 
was set if the line contained "subroutine", "function", or integer followed 
by function (why this last one??), and carried to the next lines until the 
next line containing "subroutine" was found and checked for re-flagging. 
Of course, the last subroutine, if flagged, would carry the flag until the 
end of the file. 

I added a "dear code, please set the flag to 'no write no more'" if it 
reaches "endsubroutine" or "endfunction". 

I noticed that all special/*f90 files have the last subroutine of 
nospecial.f90 in them, even if is not used. That's why this problem 
had long passed unnoticed. 

Also, added LOADS of comments (line by line) to the write_dummies routine. 
Perl is notoriously horrible to read.